bootstrap: split out components
The bootstrap components are now loaded as an import tree, which will
make it easier to test/maintain each component.
